requirements.txt
fastapi>=0.68.1
uvicorn>=0.15.0
sqlalchemy>=1.4.23
pydantic>=1.8.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
streamlit>=1.10.0
python-dotenv>=0.19.0
pandas>=1.3.3
plotly>=5.3.1
requests>=2.26.0
kiteconnect>=5.0.1 